requirement.txt
absl-py==2.1.0
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
audioread==3.0.1
babel==2.16.0
baseconfig==1.0.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.9.0
boolean.py==4.0
CacheControl==0.14.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
cyclonedx-python-lib==7.6.2
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
ecdsa==0.19.0
executing==2.1.0
fastapi==0.115.6
fastjsonschema==2.20.0
ffmpy==0.4.0
filelock==3.16.1
Flask==3.1.0
Flask-Limiter==3.9.2
flatbuffers==24.3.25
fonttools==4.55.2
fqdn==1.5.1
fsspec==2024.10.0
fvcore==0.1.5.post20221221
gradio==5.8.0
gradio_client==1.5.1
grpcio==1.68.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.26.3
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.7
inflect==7.4.0
iopath==0.1.10
ipykernel==6.29.5
ipython==8.27.0
ipython-genutils==0.2.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-highlight-selected-word==0.2.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_contrib_core==0.4.2
jupyter_contrib_nbextensions==0.7.0
jupyter_core==5.7.2
jupyter_nbextensions_configurator==0.6.4
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.10.2.post1
license-expression==30.4.0
limits==3.14.1
llvmlite==0.43.0
logging==0.4.9.6
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
modelscope==1.20.1
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.2.2
notebook_shim==0.2.4
numba==0.60.0
numpy==2.0.2
nvidia-cuda-runtime-cu12==12.6.77
onnxruntime==1.20.1
opencv-python==4.10.0.84
ordered-set==4.1.0
orjson==3.10.12
overrides==7.7.0
packageurl-python==0.16.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pillow==11.0.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.7.3
platformdirs==4.2.2
pooch==1.8.2
portalocker==3.0.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==5.29.1
psutil==6.0.0
pure_eval==0.2.3
py-serializable==1.1.2
pyasn1==0.6.1
pycocotools==2.0.8
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.10.1
pyparsing==3.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.19
pytz==2024.2
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.20.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
ruff==0.8.1
safehttpx==0.1.6
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
semantic-version==2.10.0
Send2Trash==1.8.3
setuptools==75.6.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.6
soxr==0.5.0.post1
stack-data==0.6.3
starlette==0.41.3
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorrt==10.7.0
tensorrt_cu12==10.7.0
tensorrt_cu12_bindings==10.7.0
tensorrt_cu12_libs==10.7.0
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.8.0
timm==1.0.12
tinycss2==1.3.0
tokenizers==0.20.3
toml==0.10.2
tomlkit==0.13.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.3
typeguard==4.4.1
typer==0.15.1
types-python-dateutil==2.9.0.20240821
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==2.2.2
uvicorn==0.32.1
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
whisper==1.1.10
wrapt==1.17.0
yacs==0.1.8
yamlconfig==0.3.1